Refactored super class into renderer, device and loader - Removed gltf and bin files
